DisqusCommentsActivity
Idea: have button launch activity that is a textview. A post button will
be in the top right corner. Removed some old string and codes, added the
post comment code to disqusdetails, and set up an intent to launch
disquscommentsactivity in the onclicklistener of the disqus button.
renamed disqusactivity to disqusmainactivity.