Finished refresh!, cosmetic change, bug fixes, new menu option
It was just "act.get() != null" that I forgot to check, returning an
empty list. I fixed it by only using 1 constructor and two booleans to
more clearly determine what should happen when this task is called. I
didn't want the swiperefresh animation to show because we already have a
progressbar, so I immediately set it to false after calling the task.
Sometimes the http request sometimes called a filenotfoundexception and
I wasn't sure why, so I made the empty list text give instructions for
how to refresh. Also, if you clicked the comments too quickly, before an
article loaded, the slug wouldn't be sent into the comments activity, so
I changed the button listener to only do it once the slug exists.
Lastly, there is a refresh button in the menu. Yes this is bad practice,
but it'll take time to automate everything, so this is simplest, two
ways to refresh.