Potential source of bugs found, so far no crashes
Bitmaps are set as drawable if the desired bitmap is already in the cache. This meant that there
was a place in my code that changed the drawable, but not the tag that had the image url. This was
causing an inconsistency where the zeroth index image was being recycled into the pool because the
cache didn't contain the old image url (but it did contain the zeroth bitmap). Hence, on occasion,
the zeroth bitmap appeared in both the pool and the cache, which may have led to the "using a recycled
image" glitch.