Deleted disqus and article activity related code
It should be noted that Disqus' commenting feature is at the bottom of the chrome tabs now, so there was no need for a separate
activity. Kareem's chrome tabs should be used for further web-related activities, so some web view client files have been edited
to not open an article activity. Likely, those webviewclients will be deleted as they were part of the old html scraping idea.