Progress classes to simplify view changing code in task, more simplifying
Changed arguments of fetchlistsearchtask to make it generic enough for even category code.
Made new classes to represent the complex changing of views in searchFragment without making the task
touch any of it. Further attempted to remove any searchFragment-specific reference so that fetchListSearchTask
just relies on the properties of existing classes, not specific views. It compiles and works, but logs show that
multiple requests are made, and I'm not sure why. Further inspection would be good to be sure that everything works
right. Progressbars showing up when they shouldn't. etc.