Update SearchAdapter and FetchListSearchTask for pageNum == 0
Before, we swapped the entire fragment for new searches. Another way was the DebugClear() method, but it had the bug
where it would search for an incremented page number from the adapter's current page number. I changed the SearchTask
so now, if the requested page num is 0, the adapter can simply swap its contents with the new list rather than append
the contents. This makes new searches much easier.