Moved files, minor fixes
Moved FragmentWithCache to image loading because it was more appropriate to associate it with all the other image
related classes. Minor bug fixes. Changed ProgressReactor so that success of the network request is an argument. This
way, we avoid the missing refresh instruction issue upon an unsuccessful refresh in category and search fragments.