Moved error classes to own files. Updated CategoryFragment error catching
Moved the files to their own classes because I wanted to put an enum into FetchError, and
java doesn't allow a enum to be defined in a nested class. Changed how CategoryFragment handles
errors so that it distinguishes between a network error and running out of results.

Temporarily made the adview invisible. I'll fix that next commit.